* Refactory/chatbot
1. 패키지명 변경
2. 쓰임없는 파일 정리.
3. 쓰임없는 import 삭제
